WW-4996 Renames and moves class to proper package to better expose its meaning
